+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TankAutonomous.java
@@ -8,7 +8,7 @@ import com.qualcomm.robotcore.hardware.Servo;
 
 @Autonomous(name = "TankAutonomous", group="Skystone")
 public class TankAutonomous extends LinearOpMode {
-    private AutoMethods auto = new AutoMethods("tank");
+    private TeleOpMethods robot = new TeleOpMethods();
     final  MecanumWheels mecanumWheels=new MecanumWheels("tank");
     final double HIGH_POWER = 1.0;
     final double NORMAL_POWER = 0.5;
@@ -17,13 +17,11 @@ public class TankAutonomous extends LinearOpMode {
     Servo hookServo;
 
     public void runOpMode() {
-        mecanumWheels.InitializeHardware(this);
+        robot.InitializeHardware(this);
         //robot.InitializeIMU();
 
         hookServo=hardwareMap.servo.get("hookServo");
 
-
-        //TODO remove lines 27-34 the motors are initialized with a method
         DcMotor frontRight = hardwareMap.dcMotor.get("frontRight");
         frontRight.setDirection(DcMotorSimple.Direction.REVERSE);
 
@@ -34,10 +32,9 @@ public class TankAutonomous extends LinearOpMode {
         DcMotor backLeft = hardwareMap.dcMotor.get("backLeft");
 
         mecanumWheels.initialize(frontLeft, frontRight, backLeft, backRight);
-        auto.setZeroPowerBrakeBehavior();
-
+        mecanumWheels.setZeroPowerBrakeBehavior();
         waitForStart();
-        auto.startTime();
+        robot.startTime();
 
         //robot.moveHook("up");
 
@@ -47,12 +44,16 @@ public class TankAutonomous extends LinearOpMode {
 
     }
 
-    //TODO create this method above the loop?
+
     protected void executeAutonomousLogic() {
         double ticksToInches=288/(Math.PI*6.125);
-        auto.ForwardMoveInches(telemetry, NORMAL_POWER, 20, ticksToInches);
+        mecanumWheels.ForwardMoveInches(telemetry, NORMAL_POWER, 20, ticksToInches);
+
+
 
     }
+
+
 
 
 }
