+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TankTeleOp.java
@@ -8,10 +8,11 @@ import com.qualcomm.robotcore.hardware.Servo;
 
 
 
+@SuppressWarnings("StatementWithEmptyBody")
 @TeleOp(name="Tank TeleOp", group="Skystone")
 //@Disabled
 public class TankTeleOp extends OpMode {
-    public TeleOpMethods robot = new TeleOpMethods("tank");
+    private TeleOpMethods robot = new TeleOpMethods();
     final  MecanumWheels mecanumWheels=new MecanumWheels("tank");
     private ButtonOneShot reverseButtonLogic = new ButtonOneShot();
     private ButtonOneShot powerChangeButtonLogic = new ButtonOneShot();
@@ -33,7 +34,7 @@ public class TankTeleOp extends OpMode {
         //attaching configuration names to each motor; each one of these names must match the name
         //of the motor in the configuration profile on the phone (spaces and capitalization matter)
         //or else an error will occur
-        mecanumWheels.InitializeHardware(this);
+        robot.InitializeHardware(this);
 
         hookServo=hardwareMap.servo.get("hookServo");
