+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOpMethods.java
@@ -8,8 +8,7 @@ import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
-public class TeleOpMethods {
-    //TODO phase out. this class is irrelevent save it for teleop specific code
+public class TeleOpMethods extends BasicRobotMethods {
     DcMotor frontLeft;
     DcMotor frontRight;
     DcMotor backLeft;
@@ -23,10 +22,42 @@ public class TeleOpMethods {
     final double WHEEL_DIAMETER = 6;
     final int NR40_PPR = 1120;
     final double DRIVE_WHEEL_GEAR_RATIO = 1;
-    String Chassis;
 
-    TeleOpMethods (String chassisName) {
-        Chassis = chassisName.toLowerCase();
+    @Override
+    public void InitializeHardware(OpMode opMode){
+        HardwareMap hardwareMap = opMode.hardwareMap;
+
+        frontRight = hardwareMap.dcMotor.get("frontRight");
+        frontRight.setDirection(DcMotorSimple.Direction.REVERSE);
+
+        frontLeft = hardwareMap.dcMotor.get("frontLeft");
+        frontLeft.setDirection(DcMotorSimple.Direction.REVERSE);
+
+        backRight = hardwareMap.dcMotor.get("backRight");
+
+        backLeft = hardwareMap.dcMotor.get("backLeft");
+
+        //spool = hardwareMap.dcMotor.get("spool");
+    }
+
+    public void setPowerVars(Gamepad gamepad1, boolean reverse) {
+        x_left = gamepad1.left_stick_x;
+        if (!reverse) {
+            x_right = gamepad1.right_stick_x;
+            y_left = -gamepad1.left_stick_y;
+        } else {
+            x_right = -gamepad1.right_stick_x;
+            y_left = gamepad1.left_stick_y;
+        }
+    }
+
+
+
+    public void setPower (double frontRightPower, double frontLeftPower, double backRightPower, double backLeftPower) {
+        frontRight.setPower(frontRightPower);
+        frontLeft.setPower(frontLeftPower);
+        backRight.setPower(backRightPower);
+        backLeft.setPower(backLeftPower);
     }
 
     public String reverseSense(boolean reverse) {
